requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
flask>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
DrissionPage>=4.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
playwright>=1.40.0

# Authentication & Database
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-Limiter==3.5.0
passlib==1.7.4
bcrypt>=4.0.0  # Required by passlib
pyotp==2.9.0
qrcode==7.4.2
Pillow>=10.0.0  # Required by qrcode
pytest-flask==1.3.0  # For testing
